perf: optimize terminal rendering by isolating clock component
- Extracted clock logic into `TerminalClock` component within `Terminal.tsx`
- Moved `getFormattedTime` utility to `lib/utils.ts` for shared access
- Removed unnecessary state updates in `Terminalcomp`, preventing re-renders every second
- Verified performance improvement with Playwright script (0 re-renders during idle vs ~2/sec previously)
